requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-ckeditor==6.7.1
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-shortuuidfield==0.1.3
django-taggit==6.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
